requirements.txt
aniso8601==8.0.0
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
bson==0.5.10
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==5.6.3
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mccabe==0.6.1
mongo==0.2.0
passlib==1.7.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pykerberos==1.2.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.22
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.1.0
